Refactor : 날짜 계산하는 calcDate 함수 module 분리
module 분리 후 각 js 파일에 import
