+++ public/javascripts/discover.js
@@ -1,3 +1,5 @@
+import calcDate from './utils/calc-date.js';
+
 const DiscoverHandler = class {
   constructor() {
     this.dateDomArr = document.getElementsByClassName('post-created-time');
@@ -5,39 +7,11 @@ const DiscoverHandler = class {
     this.trendTabBtn = document.getElementById('discover-order-trend');
     this.recentTabBtn = document.getElementById('discover-order-recent');
   }
-  
-  calcDate(postDate) {
-    const parsedDate = Date.parse(postDate);
-    const nowDate = Date.now();
-    const gapByMinute = Math.floor((nowDate - parsedDate)/(1000*60));
-
-    if (gapByMinute < 60) {
-      return `${gapByMinute} m`
-    }
-
-    const gapByHours = Math.floor(gapByMinute / 60);
-
-    if (gapByHours < 24) {
-      return `${gapByHours} h`
-    }
-
-    const gapByDay = Math.floor(gapByHours / 24);
-
-    if (gapByDay < 7) {
-      return `${gapByDay} d`
-    }
-
-    const gapByWeek = Math.floor(gapByDay / 7);
-
-    if (gapByWeek < 4) {
-      return `${gapByWeek} w` 
-    }
-  }
 
   displayTime() {
     const gapTimeArr = [];
     Array.from(this.dateDomArr).forEach((dom) => {
-      gapTimeArr.push(this.calcDate(dom.value));
+      gapTimeArr.push(calcDate(dom.value));
     })
 
     Array.from(this.displayDateDomArr).forEach((dom, index) => {
@@ -111,7 +85,6 @@ const DiscoverHandler = class {
       })
     }
   }
-  
 
   async followEvent(event) {
     const targetBtn = event.target;
