+++ public/javascripts/profile.js
@@ -1,42 +1,16 @@
+import calcDate from './utils/calc-date.js';
+
 const ProfileHandler = class {
   constructor() {
     this.dateDomArr = document.getElementsByClassName('post-created-time');
     this.displayDateDomArr = document.getElementsByClassName('box-header-time');
     this.profileBackground = document.getElementById('profile-info-wrapper');
   }
-  
-  calcDate(postDate) {
-    const parsedDate = Date.parse(postDate);
-    const nowDate = Date.now();
-    const gapByMinute = Math.floor((nowDate - parsedDate)/(1000*60));
-
-    if (gapByMinute < 60) {
-      return `${gapByMinute} m`
-    }
-
-    const gapByHours = Math.floor(gapByMinute / 60);
-
-    if (gapByHours < 24) {
-      return `${gapByHours} h`
-    }
-
-    const gapByDay = Math.floor(gapByHours / 24);
-
-    if (gapByDay < 7) {
-      return `${gapByDay} d`
-    }
-
-    const gapByWeek = Math.floor(gapByDay / 7);
-
-    if (gapByWeek < 4) {
-      return `${gapByWeek} w` 
-    }
-  }
 
   displayTime() {
     const gapTimeArr = [];
     Array.from(this.dateDomArr).forEach((dom) => {
-      gapTimeArr.push(this.calcDate(dom.value));
+      gapTimeArr.push(calcDate(dom.value));
     })
 
     Array.from(this.displayDateDomArr).forEach((dom, index) => {
@@ -50,8 +24,6 @@ const ProfileHandler = class {
       btn.addEventListener('click', async (event) => {
         const contentNumber = event.target.parentNode.firstChild.value;
         const result = await this.ajax().updateLikeStatus(contentNumber);
-
-        console.log(result);
 
         if (result === 'notLoggedIn') {
           alert('로그인이 필요한 서비스입니다');
