+++ views/content.pug
@@ -10,5 +10,5 @@ html(lang='ko')
       include component/page/post.pug
       include component/page/content.pug
   script(src='/javascripts/post.js')
-  script(src='/javascripts/content.js')
+  script(type="module" src='/javascripts/content.js')
   script(src='/javascripts/comment.js')
