+++ views/discover.pug
@@ -10,5 +10,5 @@ html(lang='ko')
       include component/page/post.pug
       include component/page/discover.pug
   script(src='/javascripts/post.js')
-  script(src='/javascripts/discover.js')
+  script(type="module" src='/javascripts/discover.js')
   script(src="/javascripts/library/isotope.pkgd.min.js")
