+++ views/following.pug
@@ -10,5 +10,5 @@ html(lang='ko')
       include component/page/post.pug
       include component/page/following.pug
   script(src='/javascripts/post.js')
-  script(src='/javascripts/following.js')
+  script(type="module" src='/javascripts/following.js')
   script(src="/javascripts/library/isotope.pkgd.min.js")
