+++ views/likes.pug
@@ -10,5 +10,5 @@ html(lang='ko')
       include component/page/post.pug
       include component/page/likes.pug
   script(src='/javascripts/post.js')
-  script(src='/javascripts/likes.js')
+  script(type="module" src='/javascripts/likes.js')
   script(src="/javascripts/library/isotope.pkgd.min.js")
