+++ views/profile.pug
@@ -10,5 +10,5 @@ html(lang='ko')
       include component/page/post.pug
       include component/page/profile.pug
   script(src='/javascripts/post.js')
-  script(src='/javascripts/profile.js')
+  script(type="module" src='/javascripts/profile.js')
   script(src="/javascripts/library/isotope.pkgd.min.js")
